Move streaming DKIM code to separate file
Move DkimSigningStream and related streaming code from sign.rs to a new
streaming.rs file, as requested by maintainer.

- Create src/dkim/streaming.rs with DkimSigningStream, SigningState,
  sign_streaming() method, and all streaming-related tests
- Update mod.rs to export from streaming module
- Clean up sign.rs by removing streaming code
